refactor: 외부 라이브러리와 관련된 Domain 패키지 변경
