+++ src/test/java/subway/presentation/PathGraphControllerTest.java
@@ -24,8 +24,8 @@ import subway.application.dto.ShortestPathInfoDto;
 import subway.application.dto.ShortestPathsDto;
 import subway.domain.fare.FareAmount;
 import subway.domain.line.Line;
-import subway.domain.path.PathEdge;
-import subway.domain.path.PathEdges;
+import subway.domain.path.graph.PathEdge;
+import subway.domain.path.graph.PathEdges;
 import subway.domain.section.Direction;
 import subway.domain.section.Distance;
 import subway.domain.station.Station;
